backend/requirements.txt
# Sistema de Triagem Inteligente - Sponte
# Dependências Python

# Framework Web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validação de dados
pydantic==2.5.0

# Cliente HTTP
httpx==0.25.2

# IA (Google Gemini)
google-generativeai==0.3.2

# Firebase (Firestore)
firebase-admin==6.4.0
google-cloud-firestore==2.13.1

# Utilitários
python-dotenv==1.0.0
python-multipart==0.0.6

# CORS
fastapi-cors==0.0.6

# JSON
orjson==3.9.10

# Logging
structlog==23.2.0

# Desenvolvimento (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
